Moving out KMS tests
